Separate lua_module code from Rust modules
This will hopefully make it easier to remove the Lua glue code once we
no longer need it.
