+++ src/world/shipparts/modules/gun.rs
@@ -1,15 +1,14 @@
 use crate::timer::Timer;
 use crate::world::types::{Location, Module, ModuleInputs, WorldEvent};
-use mlua::prelude::{LuaTable, LuaValue};
-use mlua::{Lua, Result, ToLua, UserData, UserDataMethods};
+use mlua::{Lua, UserData, UserDataMethods};
 
-fn process(orders: Vec<String>) -> bool {
+pub fn process(orders: Vec<String>) -> bool {
     orders.iter().any(|order| order == "shoot")
 }
 
-struct Gun {
-    charged: bool,
-    recharge_timer: Timer,
+pub struct Gun {
+    pub charged: bool,
+    pub recharge_timer: Timer,
 }
 
 impl Module for Gun {
@@ -50,30 +49,4 @@ impl UserData for Gun {
             },
         );
     }
-}
-
-pub fn lua_module(lua: &Lua) -> Result<LuaTable> {
-    let exports = lua.create_table()?;
-    exports.set(
-        "process",
-        lua.create_function(|_, orders: Vec<String>| Ok(process(orders)))?,
-    )?;
-
-    let metatable = lua.create_table()?;
-    metatable.set(
-        "__call",
-        lua.create_function(|lua, _: LuaValue| {
-            Gun {
-                charged: true,
-                recharge_timer: Timer {
-                    limit: 0.5,
-                    time: 0.5,
-                },
-            }
-            .to_lua(lua)
-        })?,
-    )?;
-    exports.set_metatable(Some(metatable));
-
-    Ok(exports)
 }
