+++ src/world/shipparts/modules/lua_interfaces/mod.rs
@@ -1,5 +1,3 @@
 pub mod gun;
 pub mod heal;
 pub mod missile_launcher;
-
-pub mod lua_interfaces;
